Add LEDs

+++ Golang/ebitengine/addressable-system/sprite.go
@@ -4,13 +4,13 @@ import "github.com/hajimehoshi/ebiten/v2"
 
 type Sprite struct {
 	sprite   *ebiten.Image
-	topLeftX int
-	topLeftY int
+	topLeftX float64
+	topLeftY float64
 }
 
 // NewSprite from an image that will be located with the top left coordinates
 // at (topLeftX, topLeftY).
-func NewSprite(sprite *ebiten.Image, topLeftX int, topLeftY int) *Sprite {
+func NewSprite(sprite *ebiten.Image, topLeftX float64, topLeftY float64) *Sprite {
 	return &Sprite{
 		sprite:   sprite,
 		topLeftX: topLeftX,
@@ -20,7 +20,7 @@ func NewSprite(sprite *ebiten.Image, topLeftX int, topLeftY int) *Sprite {
 
 func (s *Sprite) Draw(screen *ebiten.Image) {
 	op := &ebiten.DrawImageOptions{}
-	op.GeoM.Translate(float64(s.topLeftX), float64(s.topLeftY))
+	op.GeoM.Translate(s.topLeftX, s.topLeftY)
 
 	screen.DrawImage(s.sprite, op)
 }
